8 bit instance masks

+++ app/packages/looker/src/overlays/detection.ts
@@ -34,13 +34,13 @@ export default class DetectionOverlay<
   private static readonly intermediateCanvas: HTMLCanvasElement =
     typeof document !== "undefined" ? document.createElement("canvas") : null;
   private labelBoundingBox: BoundingBox;
-  private mask: Uint32Array;
+  private mask: Uint8Array;
   private selected: boolean;
 
   constructor(field, label) {
     super(field, label);
     if (this.label.mask) {
-      this.mask = new Uint32Array(this.label.mask.buffer);
+      this.mask = new Uint8Array(this.label.mask.buffer);
     }
   }
